back-end-python/requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-cartesia
livekit-plugins-elevenlabs
python-dotenv
aiohttp